modify the environment variables for vl model training.
